move the rest of the functionality from each() into spawn_task_tuple.hpp
